fixed for double to float msg

+++ LFU.java
@@ -96,6 +96,6 @@ public class LFU{
 	System.out.println();
     }
     public float getHitCount(){
-    	return (float)numberHits/100.0;
+    	return (float)numberHits/(float)100.0;
     }
 }
